Feat: refactor project structure
