--- cli/counters.go
@@ -1,15 +1,0 @@
-package cli
-
-type counters struct {
-	totalResources  int
-	taggedResources int
-	totalFiles      int
-	taggedFiles     int
-}
-
-func (self *counters) Add(other counters) {
-	self.totalResources += other.totalResources
-	self.taggedResources += other.taggedResources
-	self.totalFiles += other.totalFiles
-	self.taggedFiles += other.taggedFiles
-}
