+++ internal/tagging/aws.go
@@ -1,8 +1,6 @@
 package tagging
 
-import (
-	"github.com/env0/terratag/convert"
-)
+import "github.com/env0/terratag/internal/convert"
 
 func tagAutoscalingGroup(args TagBlockArgs) Result {
 	// for now, we count on it that if there's a single "tag" in the schema (unlike "tags" block),
