+++ internal/tagging/azure.go
